+++ example/lib/db/person_collection_free.dart
@@ -1,3 +1,5 @@
+import '../app.dart';
+import '../db/job_collection_free.dart';
 import 'package:webapp/wa_model_less.dart';
 import 'package:webapp/wa_ui.dart';
 import 'package:mongo_dart/mongo_dart.dart';
@@ -14,6 +16,16 @@ class PersonCollectionFree extends DBCollectionFree {
           '_id': DBFieldFree<ObjectId>(
             readonly: true,
             hideJson: true,
+          ),
+          'job_id': DBFieldFree<ObjectId?>(
+            defaultValue: null,
+            validators: [
+              FieldValidator.hasRelation(
+                collectionModel: JobCollectionFree(db: server.db),
+                relationField: '_id',
+                isRequired: true,
+              ),
+            ],
           ),
           'name': DBFieldFree<String?>(
             validators: [
@@ -56,10 +68,48 @@ class PersonCollectionFree extends DBCollectionFree {
       );
 
   @override
-  Map<String, Object?> toModel(
+  Future<Map<String, Object?>> toModel(
     Map<String, Object?> document, {
     List<String>? selectedFields,
-  }) {
-    return super.toModel(document, selectedFields: selectedFields);
+  }) async {
+    var model = await super.toModel(document, selectedFields: selectedFields);
+    return model;
+  }
+
+  Future<List<Map<String, Object?>>> getAllWithJob({
+    SelectorBuilder? selector,
+    Map<String, Object?>? filter,
+    FindOptions? findOptions,
+    String? hint,
+    int? skip,
+    Map<String, Object>? sort,
+    int? limit,
+    Map<String, Object>? hintDocument,
+    Map<String, Object>? projection,
+    Map<String, Object>? rawOptions,
+  }) async {
+    var res = await super.getAll(
+      selector: selector,
+      filter: filter,
+      findOptions: findOptions,
+      hint: hint,
+      skip: skip,
+      sort: sort,
+      limit: limit,
+      hintDocument: hintDocument,
+      projection: projection,
+      rawOptions: rawOptions,
+    );
+
+    for (var person in res) {
+      var jobId = person['job_id'];
+      if (jobId != null && jobId is ObjectId) {
+        person['job'] = await JobCollectionFree(db: db).getByOid(jobId);
+      } else {
+        person['job'] = null;
+      }
+    }
+
+    return res;
   }
 }